🔨 refactor(vue-html-to-paper.js): add support for window title and auto close This commit adds support for two new options to the vue-html-to-paper.js file. The first option is `windowTitle` which allows the user to specify a custom title for the print window. The second option is `autoClose` which is a boolean value that determines whether or not the print window should automatically close after printing. If `autoClose` is set to false, the window will remain open after printing.
🐛 fix(index.js): add autoClose option to close window after printing
✨ feat(index.js): add windowTitle option to customize window title
🎨 style(index.html): update custom window title to be grammatically correct
The autoClose option has been added to allow the window to be closed automatically after printing. The windowTitle option has been added to allow customization of the window title. The custom window title in index.html has been updated to be grammatically correct.

🔧 chore(index.js): add autoClose and windowTitle options to VueHtmlToPaper
This commit adds two new options to the VueHtmlToPaper object: autoClose and windowTitle. The autoClose option allows the user to specify whether the print window should automatically close after printing. The windowTitle option allows the user to specify a custom title for the print window. These options can be passed in as localOptions or as part of the global options object.